+++ bin/describe.ps1
@@ -28,6 +28,11 @@ $problems = 0
 $Queue = @()
 
 foreach ($m in Get-ChildItem $Dir "$App.*" -File) {
+    if ($m.Extension -notmatch "\.($ALLOWED_MANIFEST_EXTENSION_REGEX)") {
+        Write-UserMessage "Skipping $($m.Name)" -Info
+        continue
+    }
+
     try {
         $manifest = ConvertFrom-Manifest -Path $m.FullName
     } catch {